+++ spark_pipeline_framework/transformers/fhir_receiver/v2/test/fhir_receiver/test_fhir_receiver_list_streaming.py
@@ -1,4 +1,4 @@
-from os import path, makedirs
+from os import path, makedirs, environ
 from pathlib import Path
 from shutil import rmtree
 
@@ -18,14 +18,14 @@ from mockserver_client.mockserver_client import MockServerFriendlyClient
 
 @pytest.mark.parametrize("run_synchronously", [True, False])
 @pytest.mark.parametrize("use_data_streaming", [True, False])
-def test_fhir_receiver_list(
+def test_fhir_receiver_list_streaming(
     spark_session: SparkSession, run_synchronously: bool, use_data_streaming: bool
 ) -> None:
     # Arrange
     print()
     data_dir: Path = Path(__file__).parent.joinpath("./")
 
-    temp_folder = data_dir.joinpath("./temp")
+    temp_folder = data_dir.joinpath("../temp")
     if path.isdir(temp_folder):
         rmtree(temp_folder)
     makedirs(temp_folder)
@@ -34,13 +34,15 @@ def test_fhir_receiver_list(
 
     df: DataFrame = create_empty_dataframe(spark_session=spark_session)
 
-    test_name = "test_fhir_receiver_list"
+    environ["LOGLEVEL"] = "DEBUG"
+
+    test_name = "test_fhir_receiver_list_streaming"
     fhir_server_url = f"http://mock-server:1080/{test_name}/4_0_0"
     client = MockServerFriendlyClient("http://mock-server:1080")
     client.clear(test_name)
 
     load_mock_source_api_json_responses(
-        folder=data_dir.joinpath("fhir_list_calls"),
+        folder=data_dir.joinpath("fhir_list_streaming_calls"),
         mock_client=client,
         url_prefix=f"{test_name}/4_0_0/Patient",
     )
@@ -60,8 +62,10 @@ def test_fhir_receiver_list(
         ).transform(df)
 
     # Assert
+    print("Result DataFrame")
     json_df: DataFrame = df.sparkSession.read.json(str(patient_json_path))
     json_df.show()
     json_df.printSchema()
 
+    assert json_df.count() == 1
     assert json_df.select("resourceType").collect()[0][0] == "Patient"
