+++ spark_pipeline_framework/transformers/fhir_receiver/v2/test/fhir_receiver/test_fhir_receiver_streaming.py
@@ -25,7 +25,7 @@ def test_fhir_receiver_streaming(
     print()
     data_dir: Path = Path(__file__).parent.joinpath("./")
 
-    temp_folder = data_dir.joinpath("./temp")
+    temp_folder = data_dir.joinpath("../temp")
     if path.isdir(temp_folder):
         rmtree(temp_folder)
     makedirs(temp_folder)
