+++ spark_pipeline_framework/transformers/fhir_receiver/v2/test/fhir_receiver_real_server/test_async_real_fhir_server_get_patients_large.py
@@ -25,13 +25,13 @@ from spark_pipeline_framework.utilities.spark_data_frame_helpers import (
 
 @pytest.mark.parametrize("run_synchronously", [True, False])
 @pytest.mark.parametrize("use_data_streaming", [True, False])
-async def test_async_real_fhir_server_get_patients_large_with_limit(
+async def test_async_real_fhir_server_get_patients_large(
     spark_session: SparkSession, run_synchronously: bool, use_data_streaming: bool
 ) -> None:
     print()
     data_dir: Path = Path(__file__).parent.joinpath("./")
 
-    temp_folder = data_dir.joinpath("./temp")
+    temp_folder = data_dir.joinpath("../temp")
     if path.isdir(temp_folder):
         rmtree(temp_folder)
     makedirs(temp_folder)
@@ -113,7 +113,6 @@ async def test_async_real_fhir_server_get_patients_large_with_limit(
             auth_client_id=auth_client_id,
             auth_client_secret=auth_client_secret,
             use_data_streaming=use_data_streaming,
-            limit=5,
         ).transform_async(df)
 
     # Assert
@@ -121,4 +120,4 @@ async def test_async_real_fhir_server_get_patients_large_with_limit(
     json_df.show()
     json_df.printSchema()
 
-    assert json_df.count() == 5
+    assert json_df.count() == count
