+++ spark_pipeline_framework/transformers/fhir_receiver/v2/test/fhir_receiver_real_server/test_async_real_fhir_server_get_patients_large_with_limit.py
@@ -9,11 +9,9 @@ from helix_fhir_client_sdk.responses.fhir_merge_response import FhirMergeRespons
 from helix_fhir_client_sdk.utilities.fhir_helper import FhirHelper
 from helix_fhir_client_sdk.utilities.fhir_server_helpers import FhirServerHelpers
 from pyspark.sql import DataFrame, SparkSession
-from spark_fhir_schemas.r4.resources.patient import PatientSchema
 
 from spark_pipeline_framework.logger.yarn_logger import get_logger
 from spark_pipeline_framework.progress_logger.progress_logger import ProgressLogger
-from spark_pipeline_framework.transformers.fhir_reader.v1.fhir_reader import FhirReader
 from spark_pipeline_framework.transformers.fhir_receiver.v2.fhir_receiver import (
     FhirReceiver,
 )
@@ -27,19 +25,18 @@ from spark_pipeline_framework.utilities.spark_data_frame_helpers import (
 
 @pytest.mark.parametrize("run_synchronously", [True, False])
 @pytest.mark.parametrize("use_data_streaming", [True, False])
-async def test_async_real_fhir_server_get_patients_by_id_large(
+async def test_async_real_fhir_server_get_patients_large_with_limit(
     spark_session: SparkSession, run_synchronously: bool, use_data_streaming: bool
 ) -> None:
     print()
     data_dir: Path = Path(__file__).parent.joinpath("./")
 
-    temp_folder = data_dir.joinpath("./temp")
+    temp_folder = data_dir.joinpath("../temp")
     if path.isdir(temp_folder):
         rmtree(temp_folder)
     makedirs(temp_folder)
 
     patient_json_path: Path = temp_folder.joinpath("patient.json")
-    patient_id_json_path: Path = temp_folder.joinpath("patient_id.json")
 
     resource_type = "Patient"
     await FhirServerHelpers.clean_fhir_server_async(resource_type=resource_type)
@@ -105,25 +102,8 @@ async def test_async_real_fhir_server_get_patients_by_id_large(
     parameters = {"flow_name": "Test Pipeline V2", "team_name": "Data Operations"}
 
     with ProgressLogger() as progress_logger:
-        # first get the ids
         await FhirReceiver(
             server_url=fhir_server_url,
-            resource=resource_type,
-            file_path=patient_id_json_path,
-            progress_logger=progress_logger,
-            parameters=parameters,
-            run_synchronously=run_synchronously,
-            auth_well_known_url=auth_well_known_url,
-            auth_client_id=auth_client_id,
-            auth_client_secret=auth_client_secret,
-            use_data_streaming=use_data_streaming,
-            include_only_properties=["id"],
-            view="id_view",
-        ).transform_async(df)
-
-        await FhirReceiver(
-            server_url=fhir_server_url,
-            id_view="id_view",
             resource=resource_type,
             file_path=patient_json_path,
             progress_logger=progress_logger,
@@ -133,15 +113,7 @@ async def test_async_real_fhir_server_get_patients_by_id_large(
             auth_client_id=auth_client_id,
             auth_client_secret=auth_client_secret,
             use_data_streaming=use_data_streaming,
-        ).transform_async(df)
-
-        # now try to read it
-        await FhirReader(
-            file_path=patient_json_path,
-            view="patients",
-            name="fhir_reader",
-            progress_logger=progress_logger,
-            schema=PatientSchema.get_schema(),
+            limit=5,
         ).transform_async(df)
 
     # Assert
@@ -149,4 +121,4 @@ async def test_async_real_fhir_server_get_patients_by_id_large(
     json_df.show()
     json_df.printSchema()
 
-    assert json_df.count() == count
+    assert json_df.count() == 5
